refactor: 리팩토링
- 테스트 가능하도록 생성자에서 필드 주입받도록 수정
- dto 수정
- bridgeSize 값 객체로 수정
