+++ 5-period/java-bridge-main/src/main/java/bridge/domain/GameResult.java
@@ -1,4 +1,4 @@
-package bridge.domain.bridge;
+package bridge.domain;
 
 import java.util.Arrays;
 import java.util.Objects;
